+++ CHANGELOG.md
@@ -13,7 +13,7 @@
     * `AddAt(T item, int index)` - Returns `void` instead of `bool` and throws an exception when the item cannot be added
     * `AddAt(T[] items, int index)` - Returns `void` instead of `T[]` and throws an exception when all items cannot be added
   * `ILazyStackInventory<T>` and `LazyStackInventory<T>`
-    * `TryAdd(T item, int amount = 1)` - Returns `void` instead of `bool`
+    * `TryAdd(T item, int amount = 1)` - Makes `amount` optional with a default value of `1`
     * `Add(T item, int amount = 1)` - Returns `void` instead of `int` and throws an exception when all items cannot be added
     * `TryAddAt(T item, int index, int amount = 1)` - Makes `amount` optional with a default value of `1`
     * `AddAt(T item, int index, int amount = 1)` - Returns `void` instead of `int`, makes `amount` optional with a default value of `1`, and throws an exception when all items cannot be added